rename cmd
