+++ cmd/trsending/sftrrepo/models.go
