+++ cmd/trsending/sftrrepo/repo.go
